Add tests for the pixel selection
