+++ apps/explorer/templates/explorer/pixelset_detail.html
@@ -141,10 +141,7 @@
         {% include "foundation/form.html" %}
 
         <p>
-          <button
-            type="submit"
-            class="button secondary"
-          >
+          <button type="submit" class="button secondary">
             {% trans "Preview" %}
           </button>
